requirements.txt
Cartopy==0.22.0
matplotlib==3.8.1
netCDF4==1.7.2
numpy==1.24.1
onnx==1.16.2
onnxruntime_gpu==1.18.0
pandas==2.2.3
pytz==2025.1
